+++ src/Orchestra/Html/HtmlServiceProvider.php
@@ -62,7 +62,7 @@ class HtmlServiceProvider extends ServiceProvider
         });
 
         $this->app->bindShared('orchestra.form', function ($app) {
-            return new Form\Environment($app);
+            return new Form\Factory($app);
         });
     }
 
@@ -74,7 +74,7 @@ class HtmlServiceProvider extends ServiceProvider
     protected function registerOrchestraTableBuilder()
     {
         $this->app->bindShared('orchestra.table', function ($app) {
-            return new Table\Environment($app);
+            return new Table\Factory($app);
         });
     }